// types/FullScorecard.ts
import { GameMetaData } from "./GameMetaData";
import { MVPData } from "./MVPData";
import { Player } from "./Player";

export interface FullScorecard {
  id: number;
  player: Player;
  game: GameMetaData;
  team: string;
  position: string;
  survived: number;
  shots_hit: number;
  shots_fired: number;
  times_zapped: number;
  times_missiled: number;
  missile_hits: number;
  nukes_activated: number;
  nukes_detonated: number;
  nukes_canceled: number;
  medic_hits: number;
  own_medic_hits: number;
  medic_nukes: number;
  scout_rapid: number;
  life_boost: number;
  ammo_boost: number;
  lives_left: number;
  score: number;
  max_score: number;
  shots_left: number;
  penalty_count: number;
  shot_3hit: number;
  elim_other_team: number;
  team_elim: number;
  own_nuke_cancels: number;
  shot_opponent: number;
  shot_team: number;
  missiled_opponent: number;
  missiled_team: number;
  resupplies: number;
  rank: number;
  bases_destroyed: number;
  accuracy: number;
  hit_diff: number;
  mvp: number;
  mvp_details: MVPData;
  sp_earned: number;
  sp_spent: number;
  type: string;
  is_sub: boolean;
  uptime: number;
  resupply_downtime: number;
  other_downtime: number;
  shots_fired_during_rapid: number;
  shots_hit_during_rapid: number;
  shot_opponent_during_rapid: number;
  shot_team_during_rapid: number;
  times_team_missiled: number;
}